Removed some unnecessary stuff in assistance.py
Removed the !fwlist command for now, as all available versions can currently proceed through the guide, making knowing the serial number when purchasing a new console unnecessary at the moment. Additionally, removed part of the !downgrade command for the same reason.
Also added a warning to the !builds command informing users that most builds on the page linked are not intended for general audience.