Remove the "you need a good GPU"
Technically it is true that a good GPU is needed to bruteforce the movable.sed
However, the `.stock` command is usually presented to people who have never read the 11.4+ methods. Another thing is that most of the people doing Seedminer do not have a good GPU.
To present to these noobs such a message is off-putting. There is no need to make them feel that they cannot do this method.
Additionally (and perhaps most importantly), *you don't need a good GPU*. The amount of people who ask "miners" to bruteforce their movable.sed vastly outweigh the people who use their own GPU to do it.
Especially now that [Seedhelper](http://seedhelper.figgyc.uk/) exists, this ratio will only continue to increase.

What I want to say is that there is no need to mention this at all, because even if you don't have a good GPU, you **will** find someone who will bruteforce it for you.