Refactor the results module somewhat
-3DS is split into two files: support and results.
"support" means errors in format XXX-YYYY, which have no real relation
to the hexadecimal result codes. "results" is the old ctr.py file,
without the XXX-YYYY errors. It makes sense to separate them now because
of this lack of a relationship.

-Added hexinfo command to display
mod/summary/level/desc of input 3DS hex errors; this makes up for the
error info embeds not displaying this information

-Renamed ResultCode to ResultInfo since it's more than just a "code"

-In the near future Wii U will also be split since it follows the same
two systems as 3DS

Make linter happy