11.14 update to ntrboot command
a simple update to the .ntrboot command for the Nintendo homebrew discord server. putting a more updated picture of compatible flashcards (as well as an easier to understand layout). i also added a description with a link to 3ds.hacks.guide/ntrboot so that people can go there to see a more full list of compatible (and incompatible) flashcards for ntrboot. it also helps with figuring out where to buy each flashcart science 3ds.hacks.guide has links on where to buy each flashcart.