Implement Markdown command system
This contains the first implementation of the Markdown command system,
designed for the assistance module. It creates commands based on
Markdown files which are easily editable compared to Python commands.

Command files contain the command name and the relevant console. An
example is `ntrstream.3ds.md`. A command can have multiple files for
different consoles, such as `update.3ds.md` and `update.switch.md`.

If one console exists for the command, it will always return the
contents for that file. If multiple consoles exist, it uses the current
system for checking which file to print (based on channel name or
requested consoles).

A Markdown file contains a header with these fields (all optional):
* `title`
* `url`
* `author.name` (if setting author fields, this is required)
* `author.url`
* `author.icon-url`
* `help-desc` (shows in `.help`, last one to be loaded is the one
  actually used)
* `aliases` (merged from all files for the command)
* `color` (RGB hex code, defaults to console color)
* `thumbnail-url`

A few commands were ported to this new system as a test. This includes:
* `dsp.3ds.md`
* `nospace.3ds.md`
* `ntrstream.3ds.md`
* `update.3ds.md`
* `update.switch.md`
* `tutorial/layeredfs.3ds.md`

This still needs:
- [ ] using format_map to subtitute things like software versions
- [ ] documentation
- [ ] per-command cooldown