extras: delete togglechannel confirmation message after 10s
Given that ephemeral is True, this implies we really want to hide this
command as much as possible.

Since ephemeral doesn't work with not-app-commands, instead we just
delete the message after 10s, same behaviour as the "please do this
in #bot-cmds" message.

It seems combining the args actually works, so here goes.

The behaviour is now as follows, tested:
- app command: ephemeral, static
- command: yeets itself after 10s