Add a 3DSDB TID parser
The goal of this module is to find the Title ID of a certain game,
which can be somewhat difficult to find.

hax0kartik has 3DSDB: https://hax0kartik.github.io/3dsdb/
and also has a JSON version:
  https://github.com/hax0kartik/3dsdb/tree/master/jsons

Load them in memory, and loop every two hours (to be quite honest if
this can be done weekly, I would, but Discord doesn't seem to have
that. We don't need to do it that often, as even Kartik only checks it
daily.)

The searcher has two ways to search the list:
- By name:
  - This uses the existing Levenshtein module used to parse messages as
    a filter. The same idea applies here. This was taken from the
	previous implementation of the `unidb` command before it was moved
	to LightSage's API.
- By product code:
  - If the query sent by users is 4 letters long, we check the product
    code's last 4 letters. If they match, it will just return that. If
    it does not match, fallback to searching by game name.

This will skip all update titles by default, as that is probably
useless for most people.