Initial implementation of "title.txt" parser cog (#1405)
* Initial title.txt parser implementation

* Revert IDE mangling (sorry)

* Add DLC checking, add CMD folder check

* fuzzy match title.txt filename instead of exact match

use a regex to look for anything similar to "title.txt" instead of expecting exactly "title.txt"

* fix false-positive bad DLC detection

wrong variable was compared when checking for the presence of ".tmd" in the filename, this is now resolved

* handle helpee running tree from id1

some helpees almost get it but not quite and run "tree" from the ID1 folder instead of "title". handle this case specifically. thanks Karma

* Better handling for helpee running tree from wrong location

Will now handle being run from the root, "Nintendo 3DS", the ID0, the ID1, and "title".

* handle multiple title.txt files per message

mostly for debugging... will parse every attachment instead of terminating at the first title.txt

* fix ID1 false-positive if they've run from 00040000

wasn't checking the length of potential ID1 folders... this lead to a false positive if the helpee ran Tree from inside 00040000 somehow. this is now resolved

* revert accidental docker-compose change

* Sanitise input data to prevent @everyone pings

Input data was not previously sanitised, which meant you could escape a code-block and ping everyone. This has been resolved by replacing backticks with single quotes (backticks should NEVER appear in normal input anyway) and @ symbols with full-width versions that don't trigger Discord pings

* fix variable name in sanitize function to prevent built-in redefine

* Improve sanitisation, deduplicate message header code

* add mangled-structure detection

* Pass flake8 checks

---------

Co-authored-by: nicl83 <14316037+nicl83@users.noreply.github.com>